Prepare app for tests
