+++ src/App.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
 import MainPage from 'containers/MainPage/MainPage';
 import './styles.scss';
+import AppProviders from './AppProviders';
 
 function App() {
   return (
-    <>
+    <AppProviders>
       <MainPage />
-    </>
+    </AppProviders>
   );
 }
